+++ CHANGELOG.md
@@ -12,6 +12,18 @@ Fix Sideway Surge animation
 
 ## 0.0.4
 
-Introduced Circle Surge animation
-Introduced Dual Surge animation
-Introduced Face Surge animation
+- Introduced Circle Surge animation
+- Introduced Dual Surge animation
+- Introduced Face Surge animation
+
+## 0.0.5
+
+Update readme
+
+## 0.0.6
+
+- Introduced Round Burst animation
+- Introduced Floral Burst animation
+- Introduced Star Burst animation
+- Introduced Waving Line animation
+- Introduced Infinity Spin animation